feat(#33): content에 storage util 적용
